Refactor to close connections
The PR in #1 would have fixed this problem for Get requests, but only because of
how the SNMP package uses the connection structure. Walk requests, though, would
have used the closed connection on requests after the first and thus been
broken.

Instead, just make a new connection each time with the UDP dialer. It
appears that the structure of this package was copied from net/http, which
may not have been a good model. In any case, we can now have a better API
if one is doing multiple requests to the same host, while maintaining the
old API.

fixes #2
closes #1